LayoutTests/imported/w3c/web-platform-tests/navigation-api/commit-behavior/after-transition-traversal-commit-new-navigation-before-commit-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: lengths differ, expected array ["committed", "finished"] length 2, got ["todo"] length 1

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: lengths differ, expected array ["committed", "finished"] length 2, got ["todo"] length 1

TIMEOUT Cancel a { commit: 'after-transition' } traversal before commit() by starting a new navigation Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: lengths differ, expected array ["committed", "finished"] length 2, got ["todo"] length 1

TIMEOUT Cancel a { commit: 'after-transition' } traversal before commit() by starting a new navigation Test timed out
